Add files via upload
Here are some updated MATLAB implementations of the initial exponential step size adaptation that Bob mentioned in his initial email about this Github repo.   The "ideal" version is specific to the standard Gaussian case because it relies on a point-wise, uniform-in-steps energy error upper bound that only really holds in the Gaussian case.  The other version uses a more general way to set the parameter of the underlying exponential distribution; as Bob mentioned a better implementation would use a binary search to exactly solve for the optimal time step size which is then used to specify the conditional distribution of the time step size given the state.